Adds possibility to set up custom exception formatter.
As @georgechr indicates on #1 issue, the default behavior of logger in
https://github.com/aspnet/Logging/blob/dev/src/Microsoft.Extensions.Logging.Abstractions/LoggerExtensions.cs
is to ignore the exception content itself.

From now, the AddLog4Net() method allow the user to set up, as
parameter, the formatting style Func<> that will be used when a message
with exception was logged.

But, if the user does not configure anything to format the exception,
the process takes a default one, that concatenates the State with the
Message of the given exception.

This commit resolves #1