Fixes test problems after last PR merge.
The fix itself consist on prevent the retrieve of a non-existing
configured repository from log4net library, catching the exception and
acting to create it properly.

This commit is also used to change the structure of test projects, in
order to allow an easy maintainance on future target framework test
projects.