Bugfix/observe scope factory in options (#96)
* Exposing provider options for testing

Added an internal property on the Log4NetLogger so that tests can validate options passed to it by a Log4NetProvider. Included an AssemblyInfo to include the InternalsVisibleTo() binding for tests.

* Added new tests for Log4NetProvider

Added new tests that validate the state of the Log4NetScopeFactory on the Log4NetProviderOptions that are used when creating a new logger.

* Updated tests NetCoreApp2.0 and 2.1 test

Added the new tests to the .NET Core 2.0 and 2.1 test projects

* Adding new scope factory test to .net Core app 1.x tests

Copied the new test over to .NET core 1.x test project, since there isn't a clean way to reuse the Swords.Core.Tests project for this one

* Renaming variable to avoid field conflict

Renamed the `options` function variable to `loggerOptions` so that it doesn't make the `options` field on the class.

* Updated to use scope factory from options

Updated the `CreateLogger(string)` method to use the `ScopeFactory` included in the `Log4NetProviderOptions` field on the provider.