Make log4net support IExternalScopeProvider
Supporting the ExternalScopeProvider grants performance benefits and
allows to access scopes that are by default added to all loggers (like
Activity.Current data).

A Microsoft.Extensions.Logging.ILoggerProvider can signal to the logging
infrastructure, that it supports external scopes by implementing the
interface ISupportExternalScope. Afterwards the infrastructure will
provide the logging provider with an IExternalScopeProvider
implementation, which can be used to discover the currently active
scopes.

If the logger implements this interface it will be handled differently
by the infrastructure. Previously it received the scopes somewhat like
this:

1. A call to ILogger<A>.BeginScope("Test") would result in a call to the
   global logger [see 0] of <A> in the logging infrastructure.
2. This global logger holds all scope loggers registered with the
   logging infrastructure and calls BeginScope("Test") on each one.
3. The loggers save the scope in their own AsyncLocal<?>.

This pattern requires every logger to carry their own AsyncLocal and is
therefore suboptimal for performance if many loggers are registered. So
with ISupportExternalScope the infrastructure handels it differently:

1. A call to ILogger<A>.BeginScope("Test") results in a call to the
   global logger of <A> in the logging infrastructure.
2. The global logger adds the scope to Async of the
   IExternalScopeProvider.
3. All connected loggers received the IExternalScopeProvider on
   creation, so they can query for the scopes on every log message with
   the following method:

   IExternalScopeProvider.ForEachScope<T>(Action<object, T>, T)

This pattern requires only one AsyncLocal at the ExternalScopeProvider,
meaning additional performance. Additionally the external scope provider
is used to inject certain default scopes into the Loggers. Starting with
.NET 5 the default scope provider automatically adds the data of the
current Activity to the scope of all log messages [1][2].

[0]: https://github.com/dotnet/runtime/blob/57bfe474518ab5b7cfe6bf7424a79ce3af9d6657/src/libraries/Microsoft.Extensions.Logging/src/Logger.cs
[1]: https://github.com/dotnet/runtime/issues/34305
[2]: https://github.com/dotnet/runtime/blob/b8a8c7a4c94c02d18c0224bd94108ec5887d77f2/src/libraries/Microsoft.Extensions.Logging/src/LoggerFactoryScopeProvider.cs#L35